pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
artifacts = ["eventum/www/"]

[tool.hatch.build.targets.wheel]
packages = ["eventum"]
exclude = [
    "eventum/ui/",
    "**/tests/",
]

[project.urls]
Homepage = 'https://eventum.run'
Source = 'https://github.com/eventum-generator/eventum'
Documentation = 'https://eventum.run/docs'
Changelog = 'https://github.com/eventum-generator/eventum/blob/master/CHANGELOG.md'

[project]
name = "eventum-generator"
description = "Flexible event generator"
authors = [
    {name="Nikita Reznikov", email="nikita.reznikov.public@mail.ru"}
]
license = "Apache-2.0"
readme = "README.md"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.14",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Testing :: Mocking",
    "Topic :: Software Development :: Testing :: Traffic Generation",
    "Typing :: Typed"
]
requires-python = ">=3.14"
dependencies = [
    "aioconsole>=0.8.2",
    "aiofiles>=25.1.0",
    "aiokafka>=0.12.0",

    "case-converter>=1.2.0",
    "click>=8.3.1",
    "clickhouse-connect>=0.13.0",
    "croniter>=6.0.0",
    "dateparser>=1.3.0",
    "faker>=40.5.1",
    "fastapi>=0.134.0",
    "flatten-dict>=0.4.2",
    "griffe>=2.0.0",
    "httpx>=0.28.1",

    "jinja2>=3.1.6",
    "keyring>=25.7.0",
    "keyrings-cryptfile>=1.3.9",
    "mimesis>=19.0.0",
    "msgspec>=0.20.0",
    "numpy>=2.4.2",
    "psutil>=7.2.2",
    "pwinput>=1.0.3",
    "pydantic>=2.12.5",
    "python-multipart>=0.0.22",
    "pyyaml>=6.0.3",
    "rich>=14.3.3",
    "setproctitle>=1.3.7",
    "structlog>=25.5.0",
    "tablib>=3.9.0",
    "uvicorn>=0.41.0",

    "websockets>=16.0",
]
dynamic = ["version"]

[tool.hatch.version]
path = 'eventum/__init__.py'

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "plotly>=6.5.2",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-benchmark>=5.1.0",
    "pytest-cov>=7.0.0",
    "pytest-html>=4.2.0",
    "pytest-httpx>=0.36.0",
    "requests>=2.32.5",
    "ruff>=0.15.4",
    "types-aiofiles>=25.1.0.20251011",
    "types-croniter>=6.0.0.20250809",
    "types-dateparser>=1.3.0.20260211",
    "types-psutil>=7.2.2.20260130",
    "types-pyyaml>=6.0.12.20250915",
]

[project.scripts]
eventum = "eventum.cli.commands.eventum:cli"
eventum-keyring = "eventum.cli.commands.eventum_keyring:cli"

[tool.mypy]
exclude = [
    "^test_.*\\.py$",
]
plugins = "pydantic.mypy"

[[tool.mypy.overrides]]
module = ["aiokafka", "aiokafka.*"]
ignore_missing_imports = true
follow_untyped_imports = true

[[tool.mypy.overrides]]
module = ["griffe", "griffe.*"]
ignore_missing_imports = true
follow_untyped_imports = true

[tool.pytest.ini_options]
console_output_style = "count"
addopts = [
    "-v",
    "--cov=eventum",
    "--cov-report=html"
]
filterwarnings = [
    "ignore:Parsing dates involving a day of month:DeprecationWarning:dateparser",
]
markers = [
    "integration: tests requiring Docker backend services",
    "performance: throughput measurement tests",
    "longevity: long-running stability tests",
    "scale: multi-generator saturation tests",
    "e2e: end-to-end tests via CLI subprocess",
]

[tool.coverage.report]
omit = [
    "*/models/*",
    "*/logging/config.py",
    "*/cli/splash_screen.py",
    "*/tests/*"
]

[tool.ruff]
line-length = 79

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "A001",     # builtin-variable-shadowing
    "A002",     # builtin-argument-shadowing
    "A006",     # builtin-lambda-argument-shadowing
    "ANN401",   # any-type
    "COM812",   # missing-trailing-comma
    "D105",     # undocumented-magic-method
    "D203",     # incorrect-blank-line-before-class (incompatible with D211)
    "D205",     # missing-blank-line-after-summary
    "D213",     # multi-line-summary-second-line (incompatible with D212)
    "FIX002",   # line-contains-todo
    "PERF203",  # try-except-in-loop
    "PLC0415",  # import-outside-top-level (intentional lazy imports)
    "PLW0108",  # unnecessary-lambda
    "S311",     # suspicious-non-cryptographic-random-usage
    "S701",     # jinja2-autoescape-false
    "SIM108",   # if-else-block-instead-of-if-exp
    "TC001",    # typing-only-first-party-import
    "TC002",    # typing-only-third-party-import
    "TC003",    # typing-only-standard-library-import
    "TRY400",   # error-instead-of-exception
    "UP046",    # non-pep695-generic-class
    "UP047",    # non-pep695-generic-function
]
flake8-quotes = {inline-quotes = "single", multiline-quotes = "double"}
exclude = ["**/tests/**"]

[tool.ruff.lint.per-file-ignores]
"eventum/api/routers/**/routes.py" = [
    "D103",     # undocumented-public-function
    "FAST003"   # fast-api-unused-path-parameter
]

[tool.ruff.format]
quote-style = "single"

[tool.git-cliff.changelog]
header = """
# Changelog\n
All notable changes to this project will be documented in this file.\n
"""

# https://keats.github.io/tera/docs/#introduction
body = """
{% if version %}\
    ## [{{ version | trim_start_matches(pat="v") }}] ({{ timestamp | date(format="%Y-%m-%d") }})
{% else %}\
    ## [unreleased]
{% endif %}\
{% for group, commits in commits | group_by(attribute="group") %}
    ### {{ group | striptags | trim | upper_first }}
    {% for commit in commits %}
        - {% if commit.scope %}*({{ commit.scope }})* {% endif %}\
            {% if commit.breaking %}[**breaking**] {% endif %}\
            {{ commit.message | upper_first }}\
    {% endfor %}
{% endfor %}\n
"""

footer = """
<!-- generated by git-cliff -->
"""

trim = true

postprocessors = []

[tool.git-cliff.git]
conventional_commits = true
filter_unconventional = true
split_commits = false
commit_preprocessors = []

commit_parsers = [
  { message = "^feat", group = "<!-- 0 -->🚀 Features" },
  { message = "^fix", group = "<!-- 1 -->🐛 Bug Fixes" },
  { message = "^doc", group = "<!-- 3 -->📚 Documentation" },
  { message = "^perf", group = "<!-- 4 -->⚡ Performance" },
  { message = "^refactor", group = "<!-- 2 -->🚜 Refactor" },
  { message = "^style", group = "<!-- 5 -->🎨 Styling" },
  { message = "^test", group = "<!-- 6 -->🧪 Testing" },
  { message = "^chore\\(release\\): prepare for", skip = true },
  { message = "^chore\\(deps.*\\)", skip = true },
  { message = "^chore\\(pr\\)", skip = true },
  { message = "^chore\\(pull\\)", skip = true },
  { message = "^chore|^ci", group = "<!-- 7 -->⚙️ Miscellaneous Tasks" },
  { body = ".*security", group = "<!-- 8 -->🛡️ Security" },
  { message = "^revert", group = "<!-- 9 -->◀️ Revert" },
  { message = ".*", group = "<!-- 10 -->💼 Other" },
]

filter_commits = false
topo_order = false
sort_commits = "oldest"
